requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
deepface==0.0.72
defusedxml==0.7.1
dnspython==2.2.0
entrypoints==0.4
executing==0.8.2
facebook-sdk==3.1.0
faiss-cpu==1.7.2
faiss-gpu==1.7.2
filelock==3.4.2
Flask==2.0.2
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
gdown==4.2.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.0
importlib-resources==5.4.0
incremental==21.3.0
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mtcnn==0.1.1
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
priority==1.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
Protego==0.2.0
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pymongo==4.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==6.0.1
pytz==2021.3
pyzmq==22.3.0
queuelib==1.6.2
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.9
rsa==4.8
scipy==1.8.0
Scrapy==2.5.1
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
split-folders==0.5.1
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
Twisted==22.1.0
typing-extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
zope.interface==5.4.0
